PythonVersion/requirements.txt
discord
python-dotenv
tweepy
json
requests